+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.11.2-dev
 
 * Fix reference code links inside blockquotes.
+* Migrate tests to use the expected_output package.
 
 ## 0.11.1